@@ -1,93 +1,27 @@
-import React, { useState, useEffect } from 'react';
-import twitterLogo from './assets/twitter-logo.svg';
+import { useState, useEffect } from 'react';
+import twitterLogo from 'assets/twitter-logo.svg';
 import './App.css';
-import SelectCharacter from './Components/SelectCharacter';
-import LoadingIndicator from './Components/LoadingIndicator';
-import Arena from './Components/Arena';
-import { useContract } from './hooks/useContract';
+import SelectCharacter from 'Components/SelectCharacter';
+import LoadingIndicator from 'Components/LoadingIndicator';
+import Arena from 'Components/Arena';
+import { useContract, useAccount } from 'hooks';
 
 // Constants
 const TWITTER_HANDLE = '_buildspace';
 const TWITTER_LINK = `https://twitter.com/${TWITTER_HANDLE}`;
 
 const App = () => {
-  const [currentAccount, setCurrentAccount] = useState(null);
   const [characterNFT, setCharacterNFT] = useState(null);
   const [isLoading, setIsLoading] = useState(false);
-
+  const { currentAccount, connectWallet } = useAccount();
   const { gameContract, transformCharacterData } = useContract();
-
-  const checkIfWalletIsConnected = async () => {
-    try {
-      const { ethereum } = window;
-
-      if (!ethereum) {
-        console.log('Make sure you have MetaMask!');
-        return;
-      } else {
-        console.log('We have the ethereum object', ethereum);
-        const accounts = await ethereum.request({ method: 'eth_accounts' });
-        if (accounts.length !== 0) {
-          const account = accounts[0];
-          console.log('Found an authorized account:', account);
-          setCurrentAccount(account);
-        } else {
-          console.log('No authorized account found');
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const connectWallet = async () => {
-    try {
-      const { ethereum } = window;
-
-      if (!ethereum) {
-        alert('Get MetaMask!');
-        return;
-      }
-
-      const accounts = await ethereum.request({
-        method: 'eth_requestAccounts',
-      });
-
-      console.log('Connected', accounts[0]);
-      setCurrentAccount(accounts[0]);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  const checkNetwork = async () => {
-    try {
-      if (window.ethereum.networkVersion !== '4') {
-        alert('Please connect to Rinkeby!');
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  useEffect(() => {
-    const onLoad = async () => {
-      setIsLoading();
-      await checkIfWalletIsConnected();
-      await checkNetwork();
-    };
-    window.addEventListener('load', onLoad);
-    return () => window.removeEventListener('load', onLoad);
-  }, []);
 
   useEffect(() => {
     console.log('game contract', gameContract);
     const fetchNFTMetadata = async () => {
-      console.log('Checking for Character NFT on address:', currentAccount);
+      console.log(`Checking for Character NFT on: https://testnets.opensea.io/${currentAccount}`);
 
-      const txn = await gameContract.checkIfUserHasNFT();
+      const txn = await gameContract.getUserNFT(currentAccount);
       if (txn.name) {
         console.log('User has character NFT');
         const character = transformCharacterData(txn);
@@ -112,8 +46,11 @@ const App = () => {
 
     if (!currentAccount) {
       return (
-        <button className="cta-button connect-wallet-button" onClick={connectWallet}>
-          Connect to Wallet
+        <button
+          className="text-xl font-bold m-5 py-5 px-20 rounded-xl bg-red-500 hover:bg-red-800"
+          onClick={connectWallet}
+        >
+          Connect Wallet
         </button>
       );
     } else if (currentAccount && !characterNFT) {
@@ -124,26 +61,28 @@ const App = () => {
   };
 
   return (
-    <div className="App">
-      <div className="container">
-        <div className="header-container">
-          <p className="header gradient-text">⚔️ Pokeverse ⚔️</p>
-          <p className="sub-text">Team up to protect the Pokemon world!</p>
-
-          <div className="connect-wallet-container">
-            <img src="https://wallpaperaccess.com/full/109332.jpg" alt="first img" />
+    <div className="text-center justify-center">
+      <div className="flex flex-col justify-center items-center mx-5 sm:mx-10 3xl:mx-1/10 4xl:mx-1/6 5xl:mx-1/5 mb-20 my-10">
+        <div>
+          <div className="p-4 mb-5">
+            <p className="text-6xl font-bold mb-2">⚔️ Pokeverse ⚔️</p>
+            <p className="text-2xl italic font-bold mb-2">Team up to protect the Pokemon world!</p>
+            <p className="text-2xl font-bold">(╯°□°)╯︵◓</p>
           </div>
+          <img src="https://wallpaperaccess.com/full/109332.jpg" alt="first img" />
+
           {renderContent()}
         </div>
-        <div className="footer-container">
-          <img alt="Twitter Logo" className="twitter-logo" src={twitterLogo} />
-          <a
-            className="footer-text"
-            href={TWITTER_LINK}
-            target="_blank"
-            rel="noreferrer"
-          >{`built with @${TWITTER_HANDLE}`}</a>
-        </div>
+      </div>
+
+      <div className="flex justify-center items-center">
+        <img alt="Twitter Logo" className="w-9 h-9" src={twitterLogo} />
+        <a
+          className="text-sm font-bold underline"
+          href={TWITTER_LINK}
+          target="_blank"
+          rel="noreferrer"
+        >{`built with @${TWITTER_HANDLE}`}</a>
       </div>
     </div>
   );
