@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { abi } from '../utils/MyEpicGame.json';
+import MyEpicGame from 'contract/MyEpicGame.json';
 import { ethers } from 'ethers';
 
 export const useContract = () => {
@@ -7,7 +7,11 @@ export const useContract = () => {
   // const CONTRACT_ADDRESS = '0x047A42ef33E44628D6149023324BB82fb9f14d6a';
 
   // v2
-  const CONTRACT_ADDRESS = '0x208F0A6500DDf63E959E2B933C16Bdf7e19e26D9';
+  // const CONTRACT_ADDRESS = '0x208F0A6500DDf63E959E2B933C16Bdf7e19e26D9';
+
+  // v3
+  const CONTRACT_ADDRESS = '0xDA22e87906b085A99256554E7B3F4cB64B985c86';
+
   const [gameContract, setGameContract] = useState(null);
 
   useEffect(() => {
@@ -15,7 +19,7 @@ export const useContract = () => {
     if (ethereum) {
       const provider = new ethers.providers.Web3Provider(ethereum);
       const signer = provider.getSigner();
-      const gameContract = new ethers.Contract(CONTRACT_ADDRESS, abi, signer);
+      const gameContract = new ethers.Contract(CONTRACT_ADDRESS, MyEpicGame.abi, signer);
 
       setGameContract(gameContract);
     } else {
